Corrección de PR: Creado componente GetIconMarker.js y UserLocationMarker.js. Se cambio nombre de componente principal a LeafletMap.js y leafletMap.css. Agregado ubicación de ONG y correción de identado
