+++ src/Components/Contact/index.js
@@ -1,15 +1,15 @@
-import React from "react";
-import { MapView } from "./mapView/MapView";
-import ContactsList from "./ContactsList";
-import Title from "../Title/Title";
-import "../CardListStyles.css";
-const Contacts = () => {
+import React from 'react';
+import { LeafletMap } from './leafletMap/LeafletMap';
+import ContactsList from './ContactsList';
+import Title from '../Title/Title';
+import '../CardListStyles.css';
 
+const Contacts = () => {
   return (
     <div>
       <Title text="Contactos" />
-          <ContactsList />
-          <MapView />
+      <ContactsList />
+      <LeafletMap />
     </div>
   );
 };
